LayoutTests/imported/blink/fast/events/eventlisteneroptions/passive_inequality-expected.txt

CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.

PASS Passive Handler
PASS Non Passive Handler
PASS Passive Handler And Non Passive Capturing
